[Refactor] Rename and move scoped profiler info under ti.profiler
